Crate webb_standalone_runtime

Modules

api

constants

field_ops

hasher

hashing

impls

Some configurable implementations as associated type for the substrate runtime.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

runtime

signing

traits

types

utils

verifier

verifying

webb_proposals

Webb Protocol Proposals Specification

Structs

AdjustmentVariable

AnchorPalletId

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetDeposit

AssetsStringLimit

BagThresholds

BalanceToU256

BetterUnsignedThreshold

BlockHashCount

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

BridgeAccountId

Burn

CandidacyBond

ChainIdentifier

ChainType

ChildBountyValueMinimum

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

CurrencyAdapter

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DataDepositPerByte

DealWithFees

DefaultZeroElement

DepositBase

DepositDetails

DepositFactor

DesiredMembers

DesiredRunnersUp

DustReceiver

ElectionProviderBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ElectionsPhragmenPalletId

Element

EnactmentPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FastTrackVotingPeriod

GenesisConfig

GetNativeCurrencyId

HistoryLength

ImOnlineUnsignedPriority

IndexDeposit

InstantAllowed

LaunchPeriod

LeafDepositBase

LeafDepositPerByte

MaxApprovals

MaxAuthorities

MaxElectingVoters

MaxExtAmount

MaxFee

MaxKeys

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPending

MaxProposals

MaxProxies

MaxReserves

MaxSignatories

MaxTreeDepth

MaxVotes

MaximumReasonLength

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinPointsToBalance

MinerMaxLength

MinerMaxWeight

MinimumDeposit

MinimumMultiplier

MinimumPeriod

MixerPalletId

MultiPhaseUnsignedPriority

NoPreimagePostponement

NominationPoolsPalletId

NposSolution16

OffchainRandomBalancing

A source of random balance for NposSolver, which is meant to be run by the OCW election miner.

OffchainRepeat

OffendingValidatorsThreshold

OnChainSeqPhragmen

OnRuntimeUpgrade

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PostUnbondPoolsWindow

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ProposalLifetime

ProxyDepositBase

ProxyDepositFactor

RegistryStringLimit

ReportLongevity

RewardCurve

RootHistorySize

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SessionsPerEra

Setup

SignedDepositBase

SignedDepositByte

SignedPhase

SignedRewardBase

SlashDeferDuration

SpendPeriod

StakingBenchmarkingConfig

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

StringLimit

TargetBlockFullness

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TokenWrapperPalletId

TransactionByteFee

TreasuryPalletId

TreeDeposit

Two

U256ToBalance

UncleGenerations

UnsignedPhase

VAnchorPalletId

Version

VotingBondBase

VotingBondFactor

VotingPeriod

WebbMinerConfig

WrappingFeeDivider

Enums

Backend

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

Curve

Different curve types

Event

HashFunction

Hash functions for MerkleTree

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

Snark

Different curve types

StakerStatus

Indicates the initial status of the staker.

SystemCall

Contains one variant per dispatchable that can be called by an extrinsic.

Token

Ethereum ABI params.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

ElementTrait

HasherModule

InstanceHasher

InstanceVerifier

IntoAbiToken

VerifierModule

Functions

encode

Encodes vector of tokens into ABI compliant vector of bytes.

native_version

The version information used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

AdaptedBasicCurrency

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Amount

AssetId

AssetRegistry

AssetRegistryConfig

AssetTxPayment

Assets

AssetsForceOrigin

We allow root and the Relay Chain council to execute privileged asset operations.

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

BagsList

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

Bounties

Bridge

ChainId

Chain Id for Webb-encoded chain identifiers

CheckedExtrinsic

Extrinsic type that has already been checked.

ChildBounties

Council

CouncilConfig

Currencies

Democracy

DemocracyConfig

DigestItem

Digest item type.

ElectionProviderMultiPhase

Elections

ElectionsConfig

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

HasherBn254

HasherBn254Config

Header

Block header type as expected by this runtime.

Historical

ImOnline

ImOnlineConfig

Index

Index of a transaction in the chain.

Indices

IndicesConfig

LeafIndex

Index of a leaf in a merkle tree

LinkableTreeBn254

MerkleTreeBn254

MerkleTreeBn254Config

MixerBn254

MixerBn254Config

MixerVerifierBn254

MixerVerifierBn254Config

Moment

Type used for expressing timestamp.

Multiplier

Fee multiplier.

Multisig

NativeCurrency

NominationPools

Offences

Preimage

Proxy

RandomnessCollectiveFlip

ReserveIdentifier

Scheduler

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignatureBridge

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TokenWrapper

Tokens

TokensConfig

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

VAnchorBn254

VAnchorBn254Config

VAnchorHandlerBn254

VAnchorVerifier2x2Bn254

VAnchorVerifier2x2Bn254Config