Module webb_standalone_runtime::constants::constants

Other constants

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.