Trait webb_standalone_runtime::traits::merkle_tree::TreeInspector

pub trait TreeInspector<AccountId, TreeId, Element> {
    fn get_root(id: TreeId) -> Result<Element, DispatchError>;
    fn is_known_root(id: TreeId, target: Element) -> Result<bool, DispatchError>;
    fn get_default_root(id: TreeId) -> Result<Element, DispatchError>;
}

Tree trait for inspecting tree state

Required Methods

fn get_root(id: TreeId) -> Result<Element, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn is_known_root(id: TreeId, target: Element) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns `TreeDoesntExist

fn get_default_root(id: TreeId) -> Result<Element, DispatchError>

Gets the default merkle root for a tree or returns TreeDoesntExist