Trait webb_standalone_runtime::traits::mixer::MixerInspector

pub trait MixerInspector<AccountId, CurrencyId, TreeId, Element> {
    fn get_root(id: TreeId) -> Result<Element, DispatchError>;
    fn is_known_root(id: TreeId, target: Element) -> Result<bool, DispatchError>;
    fn ensure_known_root(
        id: TreeId, 
        target: Element
    ) -> Result<(), DispatchError>;
    fn is_nullifier_used(id: TreeId, nullifier: Element) -> bool;
    fn ensure_nullifier_unused(
        id: TreeId, 
        nullifier: Element
    ) -> Result<(), DispatchError>;
}

Mixer trait for inspecting mixer state

Required Methods

fn get_root(id: TreeId) -> Result<Element, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn is_known_root(id: TreeId, target: Element) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns `TreeDoesntExist

fn ensure_known_root(id: TreeId, target: Element) -> Result<(), DispatchError>

fn is_nullifier_used(id: TreeId, nullifier: Element) -> bool

Check if a nullifier has been used in a tree or returns InvalidNullifier

fn ensure_nullifier_unused(
id: TreeId,
nullifier: Element
) -> Result<(), DispatchError>