Type Definition webb_standalone_runtime::AllPalletsWithSystem

pub type AllPalletsWithSystem = (System, RandomnessCollectiveFlip, Timestamp, Babe, Authorship, Indices, Balances, TransactionPayment, AssetTxPayment, ElectionProviderMultiPhase, Staking, Session, Democracy, Council, Elections, Grandpa, Treasury, Utility, Multisig, Scheduler, Preimage, Proxy, Assets, Sudo, ImOnline, AuthorityDiscovery, Offences, Historical, Bounties, ChildBounties, BagsList, NominationPools, HasherBn254, AssetRegistry, Currencies, Tokens, TokenWrapper, TokenWrapperHandler, MixerVerifierBn254, VAnchorVerifier, MerkleTreeBn254, LinkableTreeBn254, MixerBn254, VAnchorBn254, VAnchorHandlerBn254, SignatureBridge, RelayerRegistry, KeyStorage);

All pallets included in the runtime as a nested tuple of types.