Module webb_standalone_runtime::types::runtime

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

AuraId

BabeId

Balance

Balance of an account.

BlockNumber

An index to a block.

DigestItem

Digest item type.

Hash

A hash of some data used by the chain.

Index

Index of a transaction in the chain.

Moment

Type used for expressing timestamp.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.