Trait webb_standalone_runtime::traits::linkable_tree::LinkableTreeInspector

pub trait LinkableTreeInspector<C> where
    C: LinkableTreeConfig, {
    fn get_chain_id() -> <C as LinkableTreeConfig>::ChainId;
    fn get_chain_id_type() -> <C as LinkableTreeConfig>::ChainId;
    fn get_chain_type() -> [u8; 2];
    fn is_known_root(
        id: <C as LinkableTreeConfig>::TreeId, 
        root: <C as LinkableTreeConfig>::Element
    ) -> Result<bool, DispatchError>;
    fn ensure_known_root(
        id: <C as LinkableTreeConfig>::TreeId, 
        root: <C as LinkableTreeConfig>::Element
    ) -> Result<(), DispatchError>;
    fn get_root(
        id: <C as LinkableTreeConfig>::TreeId
    ) -> Result<<C as LinkableTreeConfig>::Element, DispatchError>;
    fn get_neighbor_roots(
        id: <C as LinkableTreeConfig>::TreeId
    ) -> Result<Vec<<C as LinkableTreeConfig>::Element, Global>, DispatchError>;
    fn is_known_neighbor_root(
        id: <C as LinkableTreeConfig>::TreeId, 
        src_chain_id: <C as LinkableTreeConfig>::ChainId, 
        target: <C as LinkableTreeConfig>::Element
    ) -> Result<bool, DispatchError>;
    fn ensure_known_neighbor_roots(
        id: <C as LinkableTreeConfig>::TreeId, 
        roots: &Vec<<C as LinkableTreeConfig>::Element, Global>
    ) -> Result<(), DispatchError>;
    fn ensure_known_neighbor_root(
        id: <C as LinkableTreeConfig>::TreeId, 
        src_chain_id: <C as LinkableTreeConfig>::ChainId, 
        target: <C as LinkableTreeConfig>::Element
    ) -> Result<(), DispatchError>;
    fn has_edge(
        id: <C as LinkableTreeConfig>::TreeId, 
        src_chain_id: <C as LinkableTreeConfig>::ChainId
    ) -> bool;
    fn ensure_max_edges(
        id: <C as LinkableTreeConfig>::TreeId, 
        num_roots: usize
    ) -> Result<(), DispatchError>;
}

Trait for inspecting tree state

Required Methods

fn get_chain_id() -> <C as LinkableTreeConfig>::ChainId

fn get_chain_id_type() -> <C as LinkableTreeConfig>::ChainId

fn get_chain_type() -> [u8; 2]

fn is_known_root(
id: <C as LinkableTreeConfig>::TreeId,
root: <C as LinkableTreeConfig>::Element
) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns

fn ensure_known_root(
id: <C as LinkableTreeConfig>::TreeId,
root: <C as LinkableTreeConfig>::Element
) -> Result<(), DispatchError>

fn get_root(
id: <C as LinkableTreeConfig>::TreeId
) -> Result<<C as LinkableTreeConfig>::Element, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn get_neighbor_roots(
id: <C as LinkableTreeConfig>::TreeId
) -> Result<Vec<<C as LinkableTreeConfig>::Element, Global>, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn is_known_neighbor_root(
id: <C as LinkableTreeConfig>::TreeId,
src_chain_id: <C as LinkableTreeConfig>::ChainId,
target: <C as LinkableTreeConfig>::Element
) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns TreeDoesntExist

fn ensure_known_neighbor_roots(
id: <C as LinkableTreeConfig>::TreeId,
roots: &Vec<<C as LinkableTreeConfig>::Element, Global>
) -> Result<(), DispatchError>

Checks if each root from passed root array is in tree’s cached history or returns InvalidNeighborWithdrawRoot

fn ensure_known_neighbor_root(
id: <C as LinkableTreeConfig>::TreeId,
src_chain_id: <C as LinkableTreeConfig>::ChainId,
target: <C as LinkableTreeConfig>::Element
) -> Result<(), DispatchError>

Checks if a merkle root is in a tree’s cached history or returns InvalidNeighborWithdrawRoot

fn has_edge(
id: <C as LinkableTreeConfig>::TreeId,
src_chain_id: <C as LinkableTreeConfig>::ChainId
) -> bool

Check if this linked tree has this edge

fn ensure_max_edges(
id: <C as LinkableTreeConfig>::TreeId,
num_roots: usize
) -> Result<(), DispatchError>

Check if passed number of roots is the same as max allowed edges or returns InvalidMerkleRoots