Module webb_standalone_runtime::utils

Functions

compute_chain_id_type

Takes an (ideally u32) ChainIdentifier and a chain type and re-computes an updated chain id with the chain type prepended to it. The resulting chain id is 6 bytes long and so requires a u64 to represent it.

derive_resource_id

gets the chain id and tree id to derive resource id

element_encoder

get_typed_chain_id

takes an input of the underlying chain id and gets the actual chain id(Substrate) which is in u64

get_typed_chain_id_in_u64

takes an input of the underlying chain id(u32) and gets the actual chain id(Substrate) which is in u64

get_underlying_chain_id

gets the underlying chain id(u32) from the actual chain id(Substrate) which is in u64

parse_resource_id

Gets the resource id and parses it to tree id and chain id

truncate_and_pad

Truncate and pad 256 bit slice